<feat> : <Реализовано 90% функционала>

+++ src/main/java/chipization/controllers/AnimalController.java
@@ -1,90 +1,97 @@
 package chipization.controllers;
 
+import chipization.exceptions.EntityNotAuthorizedException;
 import chipization.model.Animal;
 import chipization.model.TypeAnimal;
 import chipization.model.dto.GetAnimalsRequest;
 import chipization.model.dto.TypeDto;
 import chipization.model.enums.AnimalGender;
 import chipization.model.enums.LifeStatus;
+import chipization.services.AccountService;
 import chipization.services.AnimalService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
+import java.util.ArrayList;
 import java.util.Collection;
 
 @RestController
 @RequestMapping("/animals")
 @RequiredArgsConstructor
+@Validated
 public class AnimalController {
 
     private final AnimalService animalService;
-
-    @GetMapping("/types/{typeId}")
-    public ResponseEntity<TypeAnimal> findById(@RequestHeader(value = "Authorization", required = false) String auth, @PathVariable Long typeId) {
-        return ResponseEntity.ok(animalService.findById(typeId));
-    }
-
-    @PostMapping("/types")
-    public ResponseEntity<TypeAnimal> create(@RequestHeader(value = "Authorization", required = false) String auth, @RequestBody TypeAnimal typeAnimal) {
-        return ResponseEntity.ok(animalService.create(typeAnimal));
-    }
-
-    @PutMapping("/types/{typeId}")
-    public ResponseEntity<TypeAnimal> update(@RequestHeader(value = "Authorization", required = false) String auth, @RequestBody TypeAnimal typeAnimal, @PathVariable Long typeId) {
-        return ResponseEntity.ok(animalService.update(typeId, typeAnimal));
-    }
-
-    @DeleteMapping("/types/{typeId}")
-    public void delete(@RequestHeader(value = "Authorization", required = false) String auth, @PathVariable Long typeId) {
-        animalService.delete(typeId);
-    }
+    private final AccountService accountService;
 
     @GetMapping("/{animalId}")
     public ResponseEntity<Animal> findAnimalById(@RequestHeader(value = "Authorization", required = false) String auth, @PathVariable Long animalId) {
+        accountService.checkAuthorization(auth);
         return ResponseEntity.ok(animalService.findAnimalById(animalId));
     }
 
     @GetMapping("/search")
-    public Collection<Animal> search(@RequestParam(required = false) LocalDateTime startDateTime,
-                                     @RequestParam(required = false) LocalDateTime endDateTime,
+    public Collection<Animal> search(@RequestHeader(value = "Authorization", required = false) String auth,
+                                     @RequestParam(required = false) OffsetDateTime startDateTime,
+                                     @RequestParam(required = false) OffsetDateTime endDateTime,
                                      @RequestParam(required = false) Integer chipperId,
                                      @RequestParam(required = false) Long chippingLocationId,
                                      @RequestParam(required = false) String lifeStatus,
                                      @RequestParam(required = false) String gender,
                                      @RequestParam(defaultValue = "0") Integer from,
                                      @RequestParam(defaultValue = "10") Integer size) {
+        accountService.checkAuthorization(auth);
         return animalService.search(GetAnimalsRequest.of(startDateTime, endDateTime, chipperId, chippingLocationId, lifeStatus, gender, from, size));
     }
 
     @PostMapping
-    public ResponseEntity<Animal> createAnimal(@RequestHeader(value = "Authorization", required = false) String auth, @RequestBody Animal animal) {
-        return ResponseEntity.ok(animalService.createAnimal(animal));
+    public ResponseEntity<Animal> createAnimal(@RequestHeader(value = "Authorization", required = false) String auth, @Valid @RequestBody Animal animal) {
+        if (auth == null) {
+            throw new EntityNotAuthorizedException("Анонимный пользователь, запрос недоступен");
+        }
+        accountService.checkAuthorization(auth);
+        return new ResponseEntity<>(animalService.createAnimal(animal), HttpStatus.CREATED);
     }
 
     @PutMapping("/{animalId}")
-    public ResponseEntity<Animal> updateAnimal(@RequestHeader(value = "Authorization", required = false) String auth, @RequestBody Animal animal, @PathVariable Long animalId) {
+    public ResponseEntity<Animal> updateAnimal(@RequestHeader(value = "Authorization", required = false) String auth, @Valid @RequestBody Animal animal, @PathVariable Long animalId) {
+        if (auth == null) {
+            throw new EntityNotAuthorizedException("Анонимный пользователь, запрос недоступен");
+        }
+        accountService.checkAuthorization(auth);
         return ResponseEntity.ok(animalService.updateAnimal(animalId, animal));
     }
 
     @DeleteMapping("/{animalId}")
     public void deleteAnimal(@RequestHeader(value = "Authorization", required = false) String auth, @PathVariable Long animalId) {
+        if (auth == null) {
+            throw new EntityNotAuthorizedException("Анонимный пользователь, запрос недоступен");
+        }
+        accountService.checkAuthorization(auth);
         animalService.deleteAnimal(animalId);
     }
 
     @PostMapping("/{animalId}/types/{typeId}")
     public ResponseEntity<Animal> addAnimalTypeToAnimal(@RequestHeader(value = "Authorization", required = false) String auth, @PathVariable Long animalId, @PathVariable Long typeId) {
-        return ResponseEntity.ok(animalService.addAnimalTypeToAnimal(animalId, typeId));
+        accountService.checkAuthorization(auth);
+        return new ResponseEntity<>(animalService.addAnimalTypeToAnimal(animalId, typeId), HttpStatus.CREATED);
     }
 
     @PutMapping("/{animalId}/types")
-    public ResponseEntity<Animal> updateAnimalType(@RequestHeader(value = "Authorization", required = false) String auth, @RequestBody TypeDto typeDto, @PathVariable Long animalId) {
+    public ResponseEntity<Animal> updateAnimalType(@RequestHeader(value = "Authorization", required = false) String auth, @Valid @RequestBody TypeDto typeDto, @PathVariable Long animalId) {
+        accountService.checkAuthorization(auth);
         return ResponseEntity.ok(animalService.updateAnimalType(animalId, typeDto));
     }
 
     @DeleteMapping("/{animalId}/types/{typeId}")
     public ResponseEntity<Animal> deleteAnimalTypeFromAnimal(@RequestHeader(value = "Authorization", required = false) String auth, @PathVariable Long animalId, @PathVariable Long typeId) {
+        accountService.checkAuthorization(auth);
         return ResponseEntity.ok(animalService.deleteAnimalTypeFromAnimal(animalId, typeId));
     }
 }
